Refactor and fix
Refactoring gravity and fixed moving up when on a fur surface!

+++ Assets/Scripts/Types/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    public void Damage(float healthDamage, float stunTime = 0f);
+    public void Damage(int healthDamage, float stunTime = 0f);
 }
